Amended test to expect manager method (#5)
This is needed because `add_filter()` is mocked by WP Mock itself,
and therefore cannot have expectations, while expectations are necessary.
WP Mock's hook expectations expect exact values for parameters, so it is
impossible to expect certain types. This made proper testing impossible,
since `attach()` is not guaranteed to hook the very same callable that was
passed to it(); only that the callable will get invoked.